+++ src/main/java/com/podio/contact/Profile.java
@@ -8,12 +8,12 @@ public class Profile extends ProfileUpdate {
 	/**
 	 * The id of the profile
 	 */
-	private int profileId;
+	private long profileId;
 
 	/**
 	 * The id of the user
 	 */
-	private Integer userId;
+	private Long userId;
 	
 	/**
 	 * The last time the user was seen
@@ -21,22 +21,22 @@ public class Profile extends ProfileUpdate {
 	private DateTime lastSeenOn;
 
 	@JsonProperty("profile_id")
-	public int getProfileId() {
+	public long getProfileId() {
 		return profileId;
 	}
 
 	@JsonProperty("profile_id")
-	public void setProfileId(int profileId) {
+	public void setProfileId(long profileId) {
 		this.profileId = profileId;
 	}
 
 	@JsonProperty("user_id")
-	public Integer getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
 
 	@JsonProperty("user_id")
-	public void setUserId(Integer userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
